requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
colorama==0.4.6
Django==5.0.6
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-filter==24.2
django-redis==5.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.5
sqlparse==0.5.0
tqdm==4.66.4
tzdata==2024.1
uritemplate==4.1.1
